communication channels to pipelines, allows for arbitrary communication to and from pipelines without impacting the Netbricks code
- manually cloning pipeline closures, avoids the Sync requirement when using Arc for closures
- thus we can now transmit Sender clones  (which are Send, but not Sync) to the pipelines

+++ framework/src/interface/port/phy_port.rs
@@ -151,12 +151,14 @@ impl PortQueue {
         }
     }
 
-    pub fn txq(&self) -> i32 {
-        self.txq
+    #[inline]
+    pub fn txq(&self) -> u16 {
+        self.txq as u16
     }
 
-    pub fn rxq(&self) -> i32 {
-        self.rxq
+    #[inline]
+    pub fn rxq(&self) -> u16 {
+        self.rxq as u16
     }
 }
